🚧 WIP: mcts training
