+++ configs/default.yaml
@@ -3,7 +3,12 @@ defaults:
   - _self_
   - exp: trivial
   - model: medium
-  - mcts: default
+
+training_mode: mcts
+
+mcts:
+  simulations: ${exp.trainer.rollouts}
+  childs: 10
 
 seed: 0
 device: auto
