--- configs/mcts/default.yaml
@@ -1,3 +1,0 @@
----
-simulations: ${exp.trainer.rollouts}
-childs: 10
