+++ src/mcts/__init__.py
@@ -1,1 +1,2 @@
 from .tree import MCTSTree
+from .trainer import MCTSTrainer
