+++ src/policy_gradient/__init__.py
@@ -1,2 +1,2 @@
 from .loss import PPOLoss
-from .trainer import Trainer
+from .trainer import PPOTrainer
